requirements.txt

boto3
parsel
aioredis
curl_cffi
python-dotenv
redis
selenium==4.5.0
asyncio
